wip: modern openGL

+++ code/include/g3dengine/g3dparticleemitter.h
@@ -47,11 +47,11 @@ typedef struct _G3DPARTICLE {
     G3DVECTOR  pos;
     G3DVECTOR  rot;
     G3DVECTOR  sca;
-    double     MVX[0x10];
+    float      MVX[0x10];
     G3DVECTOR  speed;
     G3DVECTOR  accel;
     int32_t    startAtFrame;
-    float    lifeTime; /*** warning ; must be SIGNED int32_t ***/
+    float      lifeTime; /*** warning ; must be SIGNED int32_t ***/
     float      transparency;
     G3DVECTOR  initialScaling;
 } G3DPARTICLE;
@@ -106,7 +106,7 @@ typedef struct _G3DPARTICLEEMITTER {
     uint32_t     type;
     G3DSCENE    *sce;
 
-    double       TIWMVX[0x10];
+    float        TIWMVX[0x10];
 } G3DPARTICLEEMITTER;
 
 /******************************************************************************/
